+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Breaking Changes
 * **change npm package scope from @ukis to @dlr-eoc:** this is breaking all imports and path mappings for @ukis!!! 
 
+* **@ukis/map-navigator:** is renamed to @ukis/map-tools
+
 * **@ukis/layer-control:** directive reverseForOf removed! 
 * **@ukis/map-ol:** duplicated exported type 'TGeoExtent' removed! - now exported from *@ukis/services-layers*
 * **@ukis/services-map-state:** duplicated exported type 'TGeoExtent' removed! - now exported from *@ukis/services-layers*